# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator API",
        "description": "API documentation for the News Aggregator application",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get articles with search, filtering, and pagination",
                "description": "Get articles with search, filtering, and pagination",
                "operationId": "58ab1dab4ab1de84ad4410cbaf322acc",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term to filter articles by title or content",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author",
                        "in": "query",
                        "description": "Filter articles by author",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_from",
                        "in": "query",
                        "description": "Filter articles from this date (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_to",
                        "in": "query",
                        "description": "Filter articles to this date (Y-m-d format)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of articles per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get a specific article by ID",
                "description": "Get a specific article by ID",
                "operationId": "c2cc119faee93d8705d6898d75123c0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Article not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Article not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/filters/sources": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Get unique sources for filtering",
                "description": "Get unique sources for filter dropdown",
                "operationId": "f8abf09a7678dad8fc87d22e49305e67",
                "responses": {
                    "200": {
                        "description": "List of unique sources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/filters/categories": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Get unique categories for filtering",
                "description": "Get unique categories for filter dropdown",
                "operationId": "2c1362ad6fb5c9cd0287446db51150e1",
                "responses": {
                    "200": {
                        "description": "List of unique categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/filters/authors": {
            "get": {
                "tags": [
                    "Filters"
                ],
                "summary": "Get unique authors for filtering",
                "description": "Get unique authors for filter dropdown",
                "operationId": "aee425188053cb1d69a30f9cc7b25094",
                "responses": {
                    "200": {
                        "description": "List of unique authors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "News article model",
                "properties": {
                    "id": {
                        "description": "Article ID",
                        "type": "integer"
                    },
                    "external_id": {
                        "description": "External API article ID",
                        "type": "string"
                    },
                    "title": {
                        "description": "Article title",
                        "type": "string"
                    },
                    "url": {
                        "description": "Article URL",
                        "type": "string",
                        "format": "uri"
                    },
                    "body": {
                        "description": "Article content",
                        "type": "string"
                    },
                    "source": {
                        "description": "Article source",
                        "type": "string"
                    },
                    "author": {
                        "description": "Article author",
                        "type": "string"
                    },
                    "image_url": {
                        "description": "Article image URL",
                        "type": "string",
                        "format": "uri",
                        "nullable": true
                    },
                    "category": {
                        "description": "Article category",
                        "type": "string"
                    },
                    "published_at": {
                        "description": "Article publication date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Record creation date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Record update date",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "BadRequest": {
                "description": "Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Standard 400 Bad Request response",
                                    "type": "string",
                                    "example": "Bad Request"
                                },
                                "errors": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Resource not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "description": "Standard 404 Not Found response",
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ValidationError": {
                "description": "Validation Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "description": "Standard 422 Validation Error response",
                                    "type": "string",
                                    "example": "The given data was invalid."
                                },
                                "errors": {
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "description": "Standard 500 Server Error response",
                                    "type": "string",
                                    "example": "Internal Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Articles",
            "description": "Articles"
        },
        {
            "name": "Filters",
            "description": "Filters"
        }
    ]
}